refactor code to make cache entry factory reusable across persistent cache implementations
